requirements.txt
pyserial==3.5
pytest==7.4.2
